+++ app/views/learning_histories/index.html.erb
@@ -19,7 +19,7 @@
       <th class="header2">学習日</th>
       <th class="header3">タグ</th>
       <th class="header4">復習した回数</th>
-      <th class="header5">完了</th>
+      <th class="header5">ステータス</th>
     </tr>
   </thead>
   <tbody>
