+++ app/views/quizzes/show.html.erb
@@ -1,0 +1,43 @@
+<div class='wrapper'>
+  <div class="container mt-5">
+    <h1 class="text-center mb-4"><%= "Quiz結果" %></h1>
+    <div class="row">
+      <div class="col-lg-8 offset-lg-2">
+        <div class="card mb-4">
+          <div class="card-body">
+            <h3 class="card-title"><%= @quiz.title %></h3></br>
+          </div>
+        </div>
+        <div class="card mb-4">
+          <div class="card-body">
+            <h5 class="card-title">Quizタイトル</h5>
+            <p class="card-text"><%= @quiz.title %></p>
+          </div>
+        </div>
+        <div class="card mb-4">
+          <div class="card-body">
+            <h5 class="card-title">クイズ詳細</h5>
+            <p class="card-text"><%= markdown(@quiz.content) %></p>
+          </div>
+        </div>
+        <div class="card mb-4">
+          <div class="card-body">
+            <h5 class="card-title">クイズ回答例</h5>
+            <p class="card-text"><%= markdown(@quiz.sample_answer) %></p>
+          </div>
+        </div>
+        <div class="card mb-4">
+          <div class="card-body">
+            <h5 class="card-title">あなたの回答</h5>
+            <p class="card-text"><%= markdown(@quiz.user_answer) %></p>
+          </div>
+        </div>
+        <div class="d-flex justify-content-center"
+          <p>内容を覚えましたか？</p>
+          <%= button_to 'はい', complete_learning_history_path(@quiz.learning_history), method: :patch, class: "btn btn-secondary" %>
+          <%= button_to 'いいえ', incomplete_learning_history_path(@quiz.learning_history), method: :patch, class: "btn btn-secondary" %>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
